Make Sendgrid API POST request
Make Sendgrid API POST request using the "to",
"subject", "body" values coming in from the request
payload.

The Sendgrid api token is stored as a secret with
Webtask.io

This works, the email is sent, although there is a socket
timeout which probably means that some response needs
to be sent.

Next step is to refactor.